gRPC/server-3 (StreamKafka)/server 3.1 (Producer)/requirements.txt
numpy
sympy
protobuf
grpcio
requests
certifi
opencv-python
opencv-contrib-python
confluent-kafka==1.8.2
pytz